Refactor: Rename all Airtable reference to Learning Content
