+++ api/lib/infrastructure/datasources/learning-content/datasource.js
@@ -27,14 +27,14 @@ const _DatasourcePrototype = {
     return cache.get(key, generator);
   },
 
-  async refreshAirtableCacheRecords() {
+  async refreshLearningContentCacheRecords() {
     const cacheKeyList = this.modelName;
     const results = await this._doList();
     await cache.set(cacheKeyList, results);
     return results;
   },
 
-  async refreshAirtableCacheRecord(id) {
+  async refreshLearningContentCacheRecord(id) {
     const cacheKeyList = this.modelName;
     const airtableRecord = await airtable.getRecord(this.tableName, id);
     const newEntry = this.fromAirTableObject(airtableRecord);
