+++ api/lib/infrastructure/repositories/target-profile-with-learning-content-repository.js
@@ -9,10 +9,10 @@ const Badge = require('../../domain/models/Badge');
 const BadgeCriterion = require('../../domain/models/BadgeCriterion');
 const BadgePartnerCompetence = require('../../domain/models/BadgePartnerCompetence');
 const Stage = require('../../domain/models/Stage');
-const skillDatasource = require('../../infrastructure/datasources/airtable/skill-datasource');
-const tubeDatasource = require('../../infrastructure/datasources/airtable/tube-datasource');
-const competenceDatasource = require('../../infrastructure/datasources/airtable/competence-datasource');
-const areaDatasource = require('../../infrastructure/datasources/airtable/area-datasource');
+const skillDatasource = require('../datasources/learning-content/skill-datasource');
+const tubeDatasource = require('../datasources/learning-content/tube-datasource');
+const competenceDatasource = require('../datasources/learning-content/competence-datasource');
+const areaDatasource = require('../datasources/learning-content/area-datasource');
 const { NotFoundError, TargetProfileInvalidError } = require('../../domain/errors');
 const { FRENCH_FRANCE } = require('../../domain/constants').LOCALE;
 const { getTranslatedText } = require('../../domain/services/get-translated-text');
