+++ api/tests/integration/infrastructure/repositories/target-profile-repository_test.js
@@ -3,7 +3,7 @@ const { expect, databaseBuilder, domainBuilder, catchErr, sinon, knex } = requir
 const TargetProfile = require('../../../../lib/domain/models/TargetProfile');
 const Skill = require('../../../../lib/domain/models/Skill');
 const targetProfileRepository = require('../../../../lib/infrastructure/repositories/target-profile-repository');
-const skillDatasource = require('../../../../lib/infrastructure/datasources/airtable/skill-datasource');
+const skillDatasource = require('../../../../lib/infrastructure/datasources/learning-content/skill-datasource');
 const { NotFoundError, AlreadyExistingEntityError } = require('../../../../lib/domain/errors');
 
 describe('Integration | Repository | Target-profile', () => {
