+++ api/tests/unit/infrastructure/datasources/learning-content/skill-datasource_test.js
@@ -1,7 +1,7 @@
 const _ = require('lodash');
 const { expect, sinon } = require('../../../../test-helper');
 const airtable = require('../../../../../lib/infrastructure/airtable');
-const skillDatasource = require('../../../../../lib/infrastructure/datasources/airtable/skill-datasource');
+const skillDatasource = require('../../../../../lib/infrastructure/datasources/learning-content/skill-datasource');
 const skillAirtableDataObjectFixture = require('../../../../tooling/fixtures/infrastructure/skillAirtableDataObjectFixture');
 const { skillRawAirTableFixture } = require('../../../../tooling/fixtures/infrastructure/skillRawAirTableFixture');
 const makeAirtableFake = require('../../../../tooling/airtable-builder/make-airtable-fake');
