+++ api/tests/unit/infrastructure/repositories/correction-repository_test.js
@@ -1,7 +1,7 @@
 const { expect, sinon, domainBuilder } = require('../../../test-helper');
 const correctionRepository = require('../../../../lib/infrastructure/repositories/correction-repository');
-const challengeDatasource = require('../../../../lib/infrastructure/datasources/airtable/challenge-datasource');
-const skillDatasource = require('../../../../lib/infrastructure/datasources/airtable/skill-datasource');
+const challengeDatasource = require('../../../../lib/infrastructure/datasources/learning-content/challenge-datasource');
+const skillDatasource = require('../../../../lib/infrastructure/datasources/learning-content/skill-datasource');
 const tutorialRepository = require('../../../../lib/infrastructure/repositories/tutorial-repository');
 const Correction = require('../../../../lib/domain/models/Correction');
 const Hint = require('../../../../lib/domain/models/Hint');
